Establish stdin connection immediately after attach
Formerly, the attach performed a lazy connection between the personality
and the portlayer servers for the stdin stream.  The reason is the
Swagger client will not make the initial connection unless there is data
present in the ReaderCloser given to the client.  We force the connection
by sending over an init byte sequence, which causes an immediate connection
to the portlayer.  The interaction portlayer was updated to ignore the initial byte
sequence.

Also, since I touched the attach code again, I went ahead and moved
all the attach and resize swagger related code to the _proxy.go file.

Resolves #2353
