+++ lib/apiservers/engine/backends/container_proxy.go
@@ -36,18 +36,25 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
+	"net/http"
 	"strings"
+	"sync"
 	"syscall"
 	"time"
 
+	"golang.org/x/net/context"
+
 	log "github.com/Sirupsen/logrus"
 	"github.com/google/uuid"
-	"github.com/mreiferson/go-httpclient"
 
 	"github.com/go-swagger/go-swagger/httpkit"
 	httptransport "github.com/go-swagger/go-swagger/httpkit/client"
+	strfmt "github.com/go-swagger/go-swagger/strfmt"
 	"github.com/go-swagger/go-swagger/swag"
+	"github.com/mreiferson/go-httpclient"
 
+	"github.com/docker/docker/api/types/backend"
+	derr "github.com/docker/docker/errors"
 	"github.com/docker/docker/pkg/stringid"
 	"github.com/docker/engine-api/types"
 	"github.com/docker/engine-api/types/container"
@@ -79,9 +86,12 @@ type VicContainerProxy interface {
 	AddInteractionToContainer(handle string, config types.ContainerCreateConfig) (string, error)
 	CommitContainerHandle(handle, imageID string) error
 	StreamContainerLogs(name string, out io.Writer, started chan struct{}, showTimestamps bool, followLogs bool, since int64, tailLines int64) error
-	ContainerRunning(vc *viccontainer.VicContainer) (bool, error)
+
+	IsRunning(vc *viccontainer.VicContainer) (bool, error)
 	Wait(vc *viccontainer.VicContainer, timeout time.Duration) (exitCode int32, processStatus string, containerState string, reterr error)
 	Signal(vc *viccontainer.VicContainer, sig uint64) error
+	Resize(vc *viccontainer.VicContainer, height, width int32) error
+	AttachStreams(ctx context.Context, vc *viccontainer.VicContainer, clStdin io.ReadCloser, clStdout, clStderr io.Writer, ca *backend.ContainerAttachConfig) error
 
 	Client() *client.PortLayer
 }
@@ -105,6 +115,7 @@ const (
 	attachPLAttemptDiff    time.Duration = 10 * time.Second
 	attachPLAttemptTimeout time.Duration = attachAttemptTimeout - attachPLAttemptDiff //timeout for the portlayer before ditching an attempt
 	attachRequestTimeout   time.Duration = 2 * time.Hour                              //timeout to hold onto the attach connection
+	attachStdinInitString                = "v1c#>"
 	swaggerSubstringEOF                  = "EOF"
 	forceLogType                         = "json-file" //Use in inspect to allow docker logs to work
 	annotationKeyLabels                  = "docker.labels"
@@ -409,8 +420,8 @@ func (c *ContainerProxy) StreamContainerLogs(name string, out io.Writer, started
 	return nil
 }
 
-// ContainerRunning returns true if the given container is running
-func (c *ContainerProxy) ContainerRunning(vc *viccontainer.VicContainer) (bool, error) {
+// IsRunning returns true if the given container is running
+func (c *ContainerProxy) IsRunning(vc *viccontainer.VicContainer) (bool, error) {
 	defer trace.End(trace.Begin(""))
 
 	if c.client == nil {
@@ -496,7 +507,7 @@ func (c *ContainerProxy) Signal(vc *viccontainer.VicContainer, sig uint64) error
 		return InternalServerError("Signal failed to create a portlayer client")
 	}
 
-	if running, err := c.ContainerRunning(vc); !running && err == nil {
+	if running, err := c.IsRunning(vc); !running && err == nil {
 		return fmt.Errorf("Signal %s is not running", vc.ContainerID)
 	}
 
@@ -510,7 +521,7 @@ func (c *ContainerProxy) Signal(vc *viccontainer.VicContainer, sig uint64) error
 				// Container may have received the signal but we got an error.
 				// Give it some time and check again.
 				c.Wait(vc, killWaitForExit)
-				if running, _ := c.ContainerRunning(vc); running {
+				if running, _ := c.IsRunning(vc); running {
 					err := InternalServerError(err.Error())
 					log.Infof("Container was signaled but is still running: %s", err)
 					return err
@@ -545,6 +556,148 @@ func (c *ContainerProxy) createNewAttachClientWithTimeouts(connectTimeout, respo
 	runtime.Producers["application/octet-stream"] = httpkit.ByteStreamProducer()
 
 	return plClient, transport
+}
+
+func (c *ContainerProxy) Resize(vc *viccontainer.VicContainer, height, width int32) error {
+	defer trace.End(trace.Begin(vc.ContainerID))
+
+	if c.client == nil {
+		return derr.NewErrorWithStatusCode(fmt.Errorf("ContainerProxy failed to create a portlayer client"),
+			http.StatusInternalServerError)
+	}
+
+	plResizeParam := interaction.NewContainerResizeParamsWithContext(ctx).
+		WithID(vc.ContainerID).
+		WithHeight(height).
+		WithWidth(width)
+
+	_, err := c.client.Interaction.ContainerResize(plResizeParam)
+	if err != nil {
+		if _, isa := err.(*interaction.ContainerResizeNotFound); isa {
+			return ResourceNotFoundError(vc.ContainerID, "interaction connection")
+		}
+
+		// If we get here, most likely something went wrong with the port layer API server
+		return InternalServerError(err.Error())
+	}
+
+	return nil
+}
+
+// attacheStreams takes the the hijacked connections from the calling client and attaches
+// them to the 3 streams from the portlayer's rest server.
+// clStdin, clStdout, clStderr are the hijacked connection
+func (c *ContainerProxy) AttachStreams(ctx context.Context, vc *viccontainer.VicContainer, clStdin io.ReadCloser, clStdout, clStderr io.Writer, ca *backend.ContainerAttachConfig) error {
+	defer clStdin.Close()
+
+	// Cancel will close the child connections.
+	ctx, cancel := context.WithCancel(ctx)
+
+	var wg sync.WaitGroup
+	errors := make(chan error, 3)
+
+	// For stdin, we only have a timeout for connection.  There can be a long duration before
+	// the first entry so there is no timeout for response.
+	plClient, transport := createNewAttachClientWithTimeouts(attachConnectTimeout, 0, attachAttemptTimeout)
+	defer transport.Close()
+
+	if ca.UseStdin {
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			err := copyStdIn(ctx, plClient, vc, clStdin, ca.DetachKeys)
+			if err != nil {
+				log.Errorf("container attach: stdin (%s): %s", vc.ContainerID, err.Error())
+			} else {
+				log.Infof("container attach: stdin (%s) done: %s", vc.ContainerID)
+			}
+
+			// no need to take action if we are canceled
+			// as that means error happened somewhere else
+			if ctx.Err() == context.Canceled {
+				log.Infof("returning from stdin as context canceled somewhere else")
+				return
+			}
+			cancel()
+
+			errors <- err
+		}()
+	}
+
+	if ca.UseStdout {
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			err := copyStdOut(ctx, plClient, attachAttemptTimeout, vc, clStdout)
+			if err != nil {
+				log.Errorf("container attach: stdout (%s): %s", vc.ContainerID, err.Error())
+			} else {
+				log.Infof("container attach: stdout (%s) done: %s", vc.ContainerID)
+			}
+
+			// no need to take action if we are canceled
+			// as that means error happened somewhere else
+			if ctx.Err() == context.Canceled {
+				log.Infof("returning from stdin as context canceled somewhere else")
+				return
+			}
+			cancel()
+
+			errors <- err
+		}()
+	}
+
+	if ca.UseStderr {
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			err := copyStdErr(ctx, plClient, vc, clStderr)
+			if err != nil {
+				log.Errorf("container attach: stderr (%s): %s", vc.ContainerID, err.Error())
+			} else {
+				log.Infof("container attach: stderr (%s) done: %s", vc.ContainerID)
+			}
+
+			// no need to take action if we are canceled
+			// as that means error happened somewhere else
+			if ctx.Err() == context.Canceled {
+				log.Infof("returning from stdin as context canceled somewhere else")
+				return
+			}
+			cancel()
+
+			errors <- err
+		}()
+	}
+
+	// Wait for all stream copy to exit
+	wg.Wait()
+
+	// close the channel so that we don't leak (if there is an error)/or get blocked (if there are no errors)
+	close(errors)
+
+	log.Infof("cleaned up connections to %s. Checking errors", vc.ContainerID)
+	for err := range errors {
+		if err != nil {
+			// check if we got DetachError
+			if _, ok := err.(DetachError); ok {
+				log.Infof("Detached from container detected")
+				return err
+			}
+
+			// If we get here, most likely something went wrong with the port layer API server
+			// These errors originate within the go-swagger client itself.
+			// Go-swagger returns untyped errors to us if the error is not one that we define
+			// in the swagger spec.  Even EOF.  Therefore, we must scan the error string (if there
+			// is an error string in the untyped error) for the term EOF.
+			log.Errorf("container attach error: %s", err)
+
+			return err
+		}
+	}
+
+	log.Infof("No error found. Returning nil...")
+	return nil
 }
 
 //----------
@@ -1176,3 +1329,191 @@ func labelsFromAnnotations(config *container.Config, annotations map[string]stri
 
 	return err
 }
+
+//------------------------------------
+// ContainerAttach() Utility Functions
+//------------------------------------
+
+func createNewAttachClientWithTimeouts(connectTimeout, responseTimeout, responseHeaderTimeout time.Duration) (*client.PortLayer, *httpclient.Transport) {
+	runtime := httptransport.New(PortLayerServer(), "/", []string{"http"})
+	transport := &httpclient.Transport{
+		ConnectTimeout:        connectTimeout,
+		ResponseHeaderTimeout: responseHeaderTimeout,
+		RequestTimeout:        responseTimeout,
+	}
+	runtime.Transport = transport
+
+	plClient := client.New(runtime, nil)
+	runtime.Consumers["application/octet-stream"] = httpkit.ByteStreamConsumer()
+	runtime.Producers["application/octet-stream"] = httpkit.ByteStreamProducer()
+
+	return plClient, transport
+}
+
+func copyStdIn(ctx context.Context, pl *client.PortLayer, vc *viccontainer.VicContainer, clStdin io.ReadCloser, keys []byte) error {
+	// Pipe for stdin so we can interject and watch the input streams for detach keys.
+	stdinReader, stdinWriter := io.Pipe()
+	defer stdinWriter.Close()
+
+	var detach bool
+
+	go func() {
+		defer stdinReader.Close()
+
+		// Copy the stdin from the CLI and write to a pipe.  We need to do this so we can
+		// watch the stdin stream for the detach keys.
+		var err error
+
+		// Write some init bytes into the pipe to force Swagger to make the initial
+		// call to the portlayer, prior to any user input in whatever attach client
+		// he/she is using.
+		log.Debugf("copyStdIn writing primer bytes")
+		stdinWriter.Write([]byte(attachStdinInitString))
+		if vc.Config.Tty {
+			_, err = copyEscapable(stdinWriter, clStdin, keys)
+		} else {
+			_, err = io.Copy(stdinWriter, clStdin)
+		}
+
+		if err != nil {
+			if _, ok := err.(DetachError); ok {
+				log.Infof("stdin detach detected")
+				detach = true
+			} else {
+				log.Errorf("stdin err: %s", err)
+			}
+		}
+	}()
+
+	// Swagger wants an io.reader so give it the reader pipe.  Also, the swagger call
+	// to set the stdin is synchronous so we need to run in a goroutine
+	setStdinParams := interaction.NewContainerSetStdinParamsWithContext(ctx).WithID(vc.ContainerID)
+	setStdinParams = setStdinParams.WithRawStream(stdinReader)
+
+	_, err := pl.Interaction.ContainerSetStdin(setStdinParams)
+	if vc.Config.StdinOnce && !vc.Config.Tty {
+		// Close the stdin connection.  Mimicing Docker's behavior.
+		// FIXME: If we close this stdin connection.  The portlayer does
+		// not really close stdin.  This is diff from current Docker
+		// behavior.  However, we're not sure why Docker even has this
+		// behavior where you connect to stdin on the first time only.
+		// If we really want to add this behavior, we need to add support
+		// in the tether in the portlayer.
+		log.Errorf("Attach stream has stdinOnce set.  VIC does not yet support this.")
+	}
+
+	// ignore the portlayer error when it is DetachError as that is what we should return to the caller when we detach
+	if detach {
+		return DetachError{}
+	}
+	return err
+}
+
+func copyStdOut(ctx context.Context, pl *client.PortLayer, attemptTimeout time.Duration, vc *viccontainer.VicContainer, clStdout io.Writer) error {
+	id := vc.ContainerID
+	//Calculate how much time to let portlayer attempt
+	plAttemptTimeout := attemptTimeout - attachPLAttemptDiff //assumes personality deadline longer than portlayer's deadline
+	plAttemptDeadline := time.Now().Add(plAttemptTimeout)
+	swaggerDeadline := strfmt.DateTime(plAttemptDeadline)
+	log.Debugf("* stdout portlayer deadline: %s", plAttemptDeadline.Format(time.UnixDate))
+	log.Debugf("* stdout personality deadline: %s", time.Now().Add(attemptTimeout).Format(time.UnixDate))
+
+	log.Debugf("* stdout attach start %s", time.Now().Format(time.UnixDate))
+	getStdoutParams := interaction.NewContainerGetStdoutParamsWithContext(ctx).WithID(id).WithDeadline(&swaggerDeadline)
+	_, err := pl.Interaction.ContainerGetStdout(getStdoutParams, clStdout)
+	log.Debugf("* stdout attach end %s", time.Now().Format(time.UnixDate))
+	if err != nil {
+		if _, ok := err.(*interaction.ContainerGetStdoutNotFound); ok {
+			return ResourceNotFoundError(id, "interaction connection")
+		}
+
+		return InternalServerError(err.Error())
+	}
+
+	return nil
+}
+
+func copyStdErr(ctx context.Context, pl *client.PortLayer, vc *viccontainer.VicContainer, clStderr io.Writer) error {
+	name := vc.ContainerID
+	getStderrParams := interaction.NewContainerGetStderrParamsWithContext(ctx).WithID(name)
+
+	_, err := pl.Interaction.ContainerGetStderr(getStderrParams, clStderr)
+	if err != nil {
+		if _, ok := err.(*interaction.ContainerGetStderrNotFound); ok {
+			ResourceNotFoundError(name, "interaction connection")
+		}
+
+		return InternalServerError(err.Error())
+	}
+
+	return nil
+}
+
+// FIXME: Move this function to a pkg to show it's origination from Docker once
+// we have ignore capabilities in our header-check.sh that checks for copyright
+// header.
+// Code c/c from io.Copy() modified by Docker to handle escape sequence
+// Begin
+
+// DetachError is special error which returned in case of container detach.
+type DetachError struct{}
+
+func (DetachError) Error() string {
+	return "detached from container"
+}
+
+func copyEscapable(dst io.Writer, src io.ReadCloser, keys []byte) (written int64, err error) {
+	if len(keys) == 0 {
+		// Default keys : ctrl-p ctrl-q
+		keys = []byte{16, 17}
+	}
+	buf := make([]byte, 32*1024)
+	for {
+		nr, er := src.Read(buf)
+		if nr > 0 {
+			// ---- Docker addition
+			preservBuf := []byte{}
+			for i, key := range keys {
+				preservBuf = append(preservBuf, buf[0:nr]...)
+				if nr != 1 || buf[0] != key {
+					break
+				}
+				if i == len(keys)-1 {
+					src.Close()
+					return 0, DetachError{}
+				}
+				nr, er = src.Read(buf)
+			}
+			var nw int
+			var ew error
+			if len(preservBuf) > 0 {
+				nw, ew = dst.Write(preservBuf)
+				nr = len(preservBuf)
+			} else {
+				// ---- End of docker
+				nw, ew = dst.Write(buf[0:nr])
+			}
+			if nw > 0 {
+				written += int64(nw)
+			}
+			if ew != nil {
+				err = ew
+				break
+			}
+			if nr != nw {
+				err = io.ErrShortWrite
+				break
+			}
+		}
+		if er == io.EOF {
+			break
+		}
+		if er != nil {
+			err = er
+			break
+		}
+	}
+	return written, err
+}
+
+// End
